[HumanBenchmark][Backend][SocialGraph] refactor social graph
