+++ back-end/api/server.js
@@ -1,45 +1,8 @@
 const express = require("express");
-const cors = require("cors");
-const helmet = require("helmet");
-const bcrypt = require("bcryptjs");
-const session = require("express-session");
-const { PrismaClient } = require("@prisma/client");
-const {
-  findUsers,
-  findUserById,
-  createUser,
-  updateUser,
-  createWorkout,
-  createBodyPartStat,
-  createMuscleStat,
-  getMuscleStats,
-  findUsersLeaderboard,
-  addFriend,
-  getFriends,
-  acceptFriendRequest,
-  deleteFriend,
-  getFriendRequests,
-  getMutualFriends,
-  getTopFriendRecommendations,
-  markWorkoutComplete,
-  deleteWorkout,
-} = require("./model-prisma");
-
-const prisma = new PrismaClient();
 const server = express();
-
-server.use((req, res, next) => {
-  console.log(`Request received: ${req.method} ${req.url}`);
-  next();
-});
-
-const eventsRouter = require("./events/events.js");
-const exercisesRouter = require("./exercises/exercises.js");
-
-server.use(helmet());
 server.use(express.json());
-server.use(eventsRouter);
-server.use(exercisesRouter);
+
+const cors = require("cors");
 server.use(
   cors({
     origin: "http://localhost:5173",
@@ -47,6 +10,12 @@ server.use(
   })
 );
 
+const helmet = require("helmet");
+server.use(helmet());
+
+const bcrypt = require("bcryptjs");
+
+const session = require("express-session");
 server.use(
   session({
     // initialize session middleware
@@ -60,8 +29,33 @@ server.use(
   })
 );
 
-server.get((req, res) => {
-  console.log("GET request received");
+const { PrismaClient } = require("@prisma/client");
+const prisma = new PrismaClient();
+
+const eventsRouter = require("./events/events.js");
+const exercisesRouter = require("./exercises/exercises.js");
+const socialGraphRouter = require("./socialGraph/socialGraph.js");
+server.use(eventsRouter);
+server.use(exercisesRouter);
+server.use(socialGraphRouter);
+
+const {
+  findUsers,
+  findUserById,
+  createUser,
+  updateUser,
+  createWorkout,
+  createBodyPartStat,
+  createMuscleStat,
+  getMuscleStats,
+  findUsersLeaderboard,
+  markWorkoutComplete,
+  deleteWorkout,
+} = require("./model-prisma");
+
+server.use((req, res, next) => {
+  console.log(`Request received: ${req.method} ${req.url}`);
+  next();
 });
 
 // [Get] /api/users
@@ -355,140 +349,6 @@ server.get("/api/leaderboard", async (req, res, next) => {
       res.json(users);
     } else {
       next({ status: 404, message: "No users found for the leaderboard" });
-    }
-  } catch (err) {
-    next(err);
-  }
-});
-
-//[Get] /api/users/:userId/friends
-server.get("/api/users/:id/friends", async (req, res, next) => {
-  // get friends by user ID
-  const id = Number(req.params.id);
-  try {
-    const friends = await getFriends(id);
-    if (friends && friends.length) {
-      res.json(friends);
-    } else {
-      next({ status: 404, message: `No friends found for user ID: ${id}` });
-    }
-  } catch (err) {
-    next(err);
-  }
-});
-
-//[Get] /api/users/:userId/friendRequests
-server.get("/api/users/:id/friendRequests", async (req, res, next) => {
-  const id = Number(req.params.id);
-  try {
-    const friendRequests = await getFriendRequests(id);
-    if (friendRequests && friendRequests.length) {
-      res.json(friendRequests);
-    } else {
-      next({
-        status: 404,
-        message: `No friend requests found for user ID: ${id}`,
-      });
-    }
-  } catch (err) {
-    next(err);
-  }
-});
-
-//[Post] /api/users/:userId/friends/:friendId
-server.post("/api/users/:userId/friends/:friendId", async (req, res, next) => {
-  //send friend request
-  const userId = Number(req.params.userId);
-  const friendId = Number(req.params.friendId);
-
-  try {
-    const request = await addFriend(userId, friendId);
-    res.status(201).json(request);
-  } catch (err) {
-    next(err);
-  }
-});
-
-//[Post] /api/users/:userId/friends/:friendId/accept
-server.post(
-  "/api/users/:userId/friends/:friendId/accept",
-  async (req, res, next) => {
-    //accept friend request
-
-    const userId = Number(req.params.userId);
-    const friendId = Number(req.params.friendId);
-
-    try {
-      const accepted = await acceptFriendRequest(userId, friendId);
-      res.status(201).json(accepted);
-    } catch (err) {
-      next(err);
-    }
-  }
-);
-
-//[Delete] /api/users/:userId/friends/:friendId/delete
-server.delete(
-  "/api/users/:userId/friends/:friendId/delete",
-  async (req, res, next) => {
-    //delete friend/ decline friend request
-    const userId = Number(req.params.userId);
-    const friendId = Number(req.params.friendId);
-    try {
-      const deleted = await deleteFriend(userId, friendId);
-      res.status(201).json(deleted);
-    } catch (err) {
-      next(err);
-    }
-  }
-);
-
-// [Get] /api/friendships
-server.get("/api/friendships", async (req, res, next) => {
-  try {
-    const friendships = await prisma.friendship.findMany();
-    if (friendships.length) {
-      res.json(friendships);
-    } else {
-      next({ status: 404, message: "No friendships found in the database" });
-    }
-  } catch (err) {
-    next(err);
-  }
-});
-
-//[Get] /api/users/:userId/recommendedFriends
-server.get("/api/users/:userId/recommendedFriends", async (req, res, next) => {
-  const userId = Number(req.params.userId);
-  try {
-    const recommendedFriends = await getTopFriendRecommendations(userId);
-    if (recommendedFriends && recommendedFriends.length) {
-      res.json(recommendedFriends);
-    } else {
-      next({
-        status: 404,
-        message: `No recommended friends found for user ID: ${userId}`,
-      });
-    }
-  } catch (err) {
-    next(err);
-  }
-});
-
-//[Get] /api/users/:userA/mutaualFriends/:userB
-server.get("/api/users/:userA/mutualFriends/:userB", async (req, res, next) => {
-  const userA = Number(req.params.userA);
-  const userB = Number(req.params.userB);
-
-  try {
-    const mutualFriends = await getMutualFriends(userA, userB);
-    if (mutualFriends && mutualFriends.length) {
-      res.json(mutualFriends);
-    } else {
-      next({
-        status: 404,
-        message: `No mutual friends found for userA ID: ${userA} and userB ID: ${userB}`,
-      });
     }
   } catch (err) {
     next(err);
